% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batchMetadt_neg20}
\alias{batchMetadt_neg20}
\title{Metabolite metadata to be included in the final library file}
\format{
batchMetadt_neg20

\describe{
\item{NAME}{character, metabolite name}
\item{FORMULA}{character, chemical formula}
\item{INCHIKEY}{character, inchikey recovered form PUBCHEM}
\item{SMILES}{character, canonical smiles recovered drom PUBCHEM}
\item{PRECURSORTYPE}{character, precursor type \link{M+H}+ for postive and \link{M-H}- for negative polarity}
\item{COLLISIONENERGY}{character, collision energy used for MS/MS}
\item{INSTRUMENTTYPE}{character, intrument type used for data collection}
}
}
\usage{
batchMetadt_neg20
}
\description{
This data frame contains the metabolite metadata to be included
in the final library file.
}
\details{
This specific data frame contains the metabolite information that was
captured in negative polarity at 20 eV (CE) in the Agilent 6545.
}
\keyword{datasets}
