% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batchRead_Neg20}
\alias{batchRead_Neg20}
\title{Data frame expressing the metabolite name, ROI, and file name}
\format{
batchRead_Neg20

A data frame containing 58 observations (metabolites) and 6 columns
\describe{
\item{Name}{character, metabolite name}
\item{Formula}{character, chemical formula of the metabolite}
\item{Ionizatioin_move}{character, the polarity (Positive or Negative)}
\item{min_rt}{integer, minimum retention time of metabolite elution (seconds)}
\item{max_rt}{integer, maximum retention time of metabolite elution (seconds)}
\item{File}{character, filename including the path}
}
}
\usage{
batchRead_Neg20
}
\description{
This data frame contains the basic info required to batch import multiple
files to create in-house libraries by using the
\href{https://github.com/CooperstoneLab/MS2extract}{MS2extract} package
}
\keyword{datasets}
