% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2extractDB-package.R
\docType{package}
\name{MS2extractDB-package}
\alias{MS2extractDB}
\alias{MS2extractDB-package}
\title{MS2extractDB: Raw MS/MS datafiles of authentic small molecules standards}
\description{
Raw MS/MS data of a set of authentic small molecules standards, including, phenolic acids, flavonoids, dihydrochalcones, carbohydrates. Standards were collected using an Agilent QTOF 6545 and 6546, in negative and positive polarities, and 20 and 60 eV (CE). Standars that were collected using the 6546 QTOF also contains MS/MS datq at 60 and 80 eV (CE). The raw MS/MS files will be used jointly with the MS2extract package to create in-house MS/MS libaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cooperstonelab.github.io/MS2extractDB/}
  \item Report bugs at \url{https://github.com/CooperstoneLab/MS2extractDB/issues}
}

}
\author{
\strong{Maintainer}: Cristian Quiroz-Moreno \email{cristianquirozd1997@gmail.com} (\href{https://orcid.org/0000-0002-9069-9147}{ORCID})

Authors:
\itemize{
  \item Jessica Cooperstone \email{cooperstone.1@osu.edu} (\href{https://orcid.org/0000-0001-7920-0088}{ORCID})
}

}
\keyword{internal}
