% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{read_pos60_6546}
\alias{read_pos60_6546}
\title{Metabolite information to batch read in positive polarity at 60 eV CE}
\format{
read_pos60_6546

A data frame containing 10 observations (metabolites) and 6 columns
\describe{
\item{Name}{character, metabolite name}
\item{Formula}{character, chemical formula of the metabolite}
\item{Ionizatioin_move}{character, the polarity (Positive or Negative)}
\item{min_rt}{integer, minimum retention time of metabolite elution (seconds)}
\item{max_rt}{integer, maximum retention time of metabolite elution (seconds)}
\item{File}{character, filename not including the path}
}
}
\usage{
read_pos60_6546
}
\description{
This data frame contains the basic info required to batch import multiple
files to create in-house libraries by using the
\href{https://github.com/CooperstoneLab/MS2extract}{MS2extract} package.
}
\details{
Metabolites included in this list were analyzed by targeted methods using
an Agilent 6546 QTOF.
}
\keyword{datasets}
