% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smiles2classyfire}
\alias{smiles2classyfire}
\title{Classyfire results of the metabolite list}
\format{
smiles2classyfire

\describe{
\item{Name}{metabolite name present in this package}
\item{INCHIKEY}{manually retrieved INCHIKEY}
\item{SMILES}{manually retrieved smiles}
\item{SMILES_CF}{smiles reported by classyfire}
\item{inchikey_CF}{inchikey reported by classyfire}
\item{Kingdom_CF}{Chemical classification reported by classyfire}
\item{Superclass_CF}{Chemical classification reported by classyfire}
\item{Class_CF}{Chemical classification reported by classyfire}
\item{Subclass_CF}{Chemical classification reported by classyfire}
}
}
\usage{
smiles2classyfire
}
\description{
This data frame contains the results of submitting the smiles chemical
structure to classyfire webpage \emph{(classyfire.wishartlab.com)}.
}
\details{
You can replicate this query submitted on August 27th 2023 by
exploring the \code{inst/extdata/classyfire} folder contained in this package.

In this folder, you will find the following:
\itemize{
\item \code{smiles2classyfire.tsv}: the table submitted to classyfire webpage
\item \code{query_11202230.json}: the json file retrieved from this query
\item \code{classydire_results.csv}: a tidy version of the classyfire results.
}
}
\keyword{datasets}
